Task 1: Software configuration.
Subtask 1: Why did I choose to participate in a challenge portfolio?

Hello my name is Iryna. Recently the lives of all Ukrainians changed. And I decided that need to change my life and my job too. First of all I noticed how cool is to have a job with me on a laptop and work no matter where I live.

Thank you for this opportunity to start learning something new.

I hope to find a job or intership and keep learning!

                Iryna